refactor: pr review 수정
